Add Dummy Program Files
Has the functionality to create programs and sent to database; however, functionality needs to be merged with Joseph's code. Right now its quite simple.
View program allows the fetch of the program and viewing it as a grid with checkboxes. Also allows for printing and downloading as pdf (might make final version)
Styles.css and script.js are for program.html 
Needs to be merged with the Dashboard styling (basically have the sidebar and inside the dashboard page)